changed interface params and return values from int to double
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@9052 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: be238ebe280d26e86a517157689647e95732ab22

+++ scheduling/scheduler.slascheduler/src/de/uniol/inf/is/odysseus/scheduler/slascheduler/ICostFunction.java
@@ -4,8 +4,8 @@ import de.uniol.inf.is.odysseus.scheduler.slamodel.SLA;
 
 public interface ICostFunction {
 	
-	public int oc(int conformance, SLA sla);
+	public double oc(double conformance, SLA sla);
 	
-	public int mg(int conformance, SLA sla);
+	public double mg(double conformance, SLA sla);
 
 }
