+++ scheduling/scheduler.slascheduler/src/de/uniol/inf/is/odysseus/scheduler/slascheduler/priority/PrioMax.java
@@ -5,9 +5,9 @@ import de.uniol.inf.is.odysseus.scheduler.slascheduler.IPriorityFunction;
 public class PrioMax implements IPriorityFunction {
 
 	@Override
-	public int calcPriority(int... cost) {
-		int max = 0;
-		for (int c : cost) {
+	public double calcPriority(double... cost) {
+		double max = 0;
+		for (double c : cost) {
 			if (c > max)
 				max = c;
 		}
