+++ scheduling/scheduler.slascheduler/src/de/uniol/inf/is/odysseus/scheduler/slascheduler/priority/PrioSum.java
@@ -5,9 +5,9 @@ import de.uniol.inf.is.odysseus.scheduler.slascheduler.IPriorityFunction;
 public class PrioSum implements IPriorityFunction {
 
 	@Override
-	public int calcPriority(int... cost) {
-		int sum = 0;
-		for (int c : cost) {
+	public double calcPriority(double... cost) {
+		double sum = 0;
+		for (double c : cost) {
 			sum += c;
 		}
 		return sum;
